move ratev2 register
